Created RegressionSQLRenumbering
This is used by me for a pretty specific task, but really saves me a lot of time. Will take a txt file and renumber to a new one. The imputed text file should act like a template. I use a different file for my output when using on my machine, but that is not needed.